refactoring to becoming a DDD application: creating repository, service and domain js resources
